// ttlogix-web/App/ClientApp/src/store/mutation-types.ts
export enum MutationTypes {
    SET_LIST_INV = "SET_LIST_INV",
    SET_LIST_ADJ = "SET_LIST_ADJ",
    SET_LIST_SD = "SET_LIST_SD",
    SET_ADJ_LIST = 'SET_ADJ_LIST',
    SET_ADJ_ITEMS = 'SET_ADJ_ITEMS',
    SET_NEW_ADJ = 'SET_NEW_ADJ',
    SET_CUSTOMERS = "SET_CUSTOMERS",
    SET_PAGING_DATA = 'SET_PAGING_DATA',
    SET_DICT = 'SET_DICT',
    SET_WHS = 'SET_WHS',
    UPDATE_ADJ = 'UPDATE_ADJ_ITEM',
    ADD_ADJ_ITEM = 'ADD_ADJ_ITEM',
    DEL_ADJ_ITEM = 'DEL_ADJ_ITEM',
    CANCEL_ADJ_ITEM = 'CANCEL_ADJ_ITEM',

    // Decant
    INV_DEC_SET_LIST = 'INV_DEC_SET_LIST',
    INV_DEC_UPDATE = 'INV_DEC_UPDATE',
    INV_DEC_SET_DECANT = 'INV_DEC_SET_DECANT',

    // Quarantine
    INV_QNE_SET_LIST = 'INV_QNE_SET_LIST',
    INV_QNE_UPDATE_REASON = 'INV_QNE_UPDATE_REASON',

    // Relocation
    INV_REL_SET_LIST = 'INV_REL_SET_LIST',

    // Common
    SET_LOCATIONS = 'SET_LOCATIONS',

    // Invoicing
    SET_FACTORIES = 'SET_FACTORIES',
    SET_LIST_INVC = 'SET_LIST_INVC',
    SET_INVC_DETAILS = 'SET_INVC_DETAILS',
    SET_APP_SETTINGS = 'SET_APP_SETTINGS',
    INVC_SHOW_INVOICE_PDF = 'INVC_SHOW_INVOICE_PDF',
    SET_ILOG_ENABLED = 'SET_ILOG_ENABLED',
    SET_PIDS_IN_ILOG = 'SET_PIDS_IN_ILOG',
    SET_PIDS_REQUESTED_FROM_ILOG = 'SET_PIDS_REQUESTED_FROM_ILOG',

    SET_CAN_EDIT_UNLOADING_POINT = 'SET_CAN_EDIT_UNLOADING_POINT',

    // Inbound Reversal
    SET_LIST_IR = 'SET_LIST_IR',
    SET_NEW_IR = 'SET_NEW_IR',
    COMPLETE_OR_CANCEL_IR = 'COMPLETE_OR_CANCEL_IR',
    SET_IR_ITEMS = 'SET_IR_ITEMS',
    UPDATE_IR_STATUS = 'UPDATE_IR_STATUS',
    UPDATE_IR = 'UPDATE_IR',

    // Inbound Reversal
    SET_LIST_STR = 'SET_LIST_STR',
    SET_NEW_STR = 'SET_NEW_STR',
    COMPLETE_OR_CANCEL_STR = 'COMPLETE_OR_CANCEL_STR',
    SET_STR_ITEMS = 'SET_STR_ITEMS',
    UPDATE_STR_STATUS = 'UPDATE_STR_STATUS',
    UPDATE_STR = 'UPDATE_STR',

    // Company Profiles
    SET_LIST_CP = 'SET_LIST_CP',
    UPDATE_COMPANY_PROFILE = 'UPDATE_COMPANY_PROFILE',
    UPDATE_ADDRESS_BOOK = 'UPDATE_ADDRESS_BOOK',
    UPDATE_ADDRESS_CONTACT = 'UPDATE_ADDRESS_CONTACT',

    // Customer
    SET_LIST_CS = 'SET_LIST_CS',
    UPDATE_CUSTOMER = 'UPDATE_CUSTOMER',
    SET_PRODUCT_CODES = 'SET_PRODUCT_CODES',
    SET_CONTROL_CODES = 'SET_CONTROL_CODES',
    SET_GLOBAL_UOM = 'SET_GLOBAL_UOM',

    // Stock Take
    SET_LIST_STL = 'SET_LIST_STL',
    UPDATE_STL = 'UPDATE_STL',
    // Settings
    SET_AREA_LIST = 'SET_AREA_LIST',
    SET_AREATYPE_LIST = 'SET_AREATYPE_LIST',
    SET_CC_LIST = 'SET_CC_LIST',
    SET_LOCATION_LIST = 'SET_LOCATION_LIST',
    SET_PACKAGETYPE_LIST = 'SET_PACKAGETYPE_LIST',
    SET_PC_LIST = 'SET_PC_LIST',
    SET_SATO_LIST = 'SET_SATO_LIST',
    SET_UOM_LIST = 'SET_UOM_LIST',
    SET_WAREHOUSE_LIST = 'SET_WAREHOUSE_LIST',
    UPDATE_LOCATION_SETTINGS = 'UPDATE_LOCATION_SETTINGS'
}
  