// ttlogix-web/App/ClientApp/src/store/models/quarantine.ts
export class Quarantine {
    pid: string;
    whsCode: string;
    productCode: string;
    supplierId: string;
    customerCode: string;
    customerName: string;
    qty: number;
    location: string;
    decimalNum: number;
    reason: string;
    createdBy: string;
    quarantineDate: string;

    // local attributes
    selected: boolean;
}