// ttlogix-web/App/ClientApp/src/store/models/relocation.ts
export class Relocation {
    pid: string;
    externalPID: string;
    productCode: string;
    supplierId: string;
    customerCode: string;
    customerName: string;
    qty: number;
    oldWhsCode: string;
    oldLocation: string;
    newWhsCode: string;
    newLocation: string;
    scannerType: number;
    relocatedBy: string;
    relocationDate: string;
}