// ttlogix-web/App/ClientApp/src/store/models/settings/location.ts
export class LocationSettings {
    code: string | null;
    warehouseCode: string | null;
    areaCode: string | null;
    name: string | null;
    m3: number | null;
    status: number | null;
    type: number | null;
    isPriority: number | null;
    iLogLocationCategoryId: number | null;

    // local props
    selected: boolean;
}