// ttlogix-web/App/ClientApp/src/store/models/settings/warehouse.ts
export class Warehouse {
    code: string | null;
    name: string | null;
    area: number | null;
    type: number | null;
    status: number | null;

    address1: string | null;
    address2: string | null;
    address3: string | null;
    address4: string | null;
    postCode: string | null;
    country: string | null;

    pic: string | null;
    telNo: string | null;
    faxNo: string | null;
}